pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "falcon-perception"
version = "0.1.0"
description = "Falcon Perception is a dense vision-language model that performs object detection, instance segmentation, and OCR from natural language queries."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    # Core (framework-agnostic)
    "einops>=0.8.1",
    "numpy>=1.24",
    "pillow>=12.0.0",
    "tokenizers>=0.21.0",
    "datasets>=4.5.0",
    "pycocotools>=2.0.11",
    "opencv-python>=4.12.0.88",
    "scipy>=1.15.3",
    "tyro>=1.0.3",
    "hf-transfer>=0.1.9",
    "requests>=2.31.0",
    "hf-xet>=1.3.1",
    # Auto-detect backend by platform
    "mlx>=0.30.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "torch>=2.11.0; sys_platform != 'darwin'",
    "torchvision>=0.26.0; sys_platform != 'darwin'",
    "safetensors>=0.7.0",
]

[project.optional-dependencies]
torch = [
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
mlx = [
    "mlx>=0.30.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
]
ocr = [
    "falcon-perception[torch]",
    "transformers>=4.57.3",
]
dev = [
    "tensorboard>=2.20.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
]
server = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.22",
]
demo = [
    "streamlit>=1.52.2",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["falcon_perception*"]

# ---------------------------------------------------------------------------
# uv: use the PyTorch cu128 index for torch / torchvision so we get wheels
# built against CUDA 12.8 (compatible with driver ≥ 570.x / CUDA 12.8).
# Without this, uv pulls the default PyPI wheels which since torch 2.11
# ship CUDA 13.0 and require a newer driver.
# ---------------------------------------------------------------------------
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
